refactor(fsd): add server entrypoints for entities
