+++ src/app/api/brands/route.js
@@ -1,4 +1,4 @@
-import { getAllBrands } from '@/entities/brand';
+import { getAllBrands } from '@/entities/brand/server';
 import { ErrorHandler } from '@/shared/lib';
 
 export const GET = ErrorHandler(async () => {
