+++ src/entities/brand/server.js
@@ -1,0 +1,3 @@
+import 'server-only';
+
+export { getAllBrands } from './api/server';
