+++ src/entities/product/server.js
@@ -1,0 +1,3 @@
+import 'server-only';
+
+export { getProductDetailsById, getProductsPaged } from './api/server';
