+++ src/entities/review/server.js
@@ -1,0 +1,3 @@
+import 'server-only';
+
+export { createReview, getReviewsByProduct, getReviewsFeed } from './api/server';
